group all superclasses into one file
* update circus.tests with new path to superclasses
* add example to sys.path in tests not test_models
* update model file modules, should all be same full path
* update example tests
