+++ circus/tests/__init__.py
@@ -2,16 +2,22 @@
 circus tests
 """
 
-import os
+# XXX: add pvpower to python path for circus testing!
+from circus.core import logging
 import imp
+import os
+import sys
 
-MODEL = 'PVPower'
+LOGGER = logging.getLogger(__name__)
+PROJECT = 'PVPower'
+MODEL = 'sandia_performance_model'
 TESTS_DIR = os.path.abspath(os.path.dirname(__file__))
 PROJ_PATH = os.path.abspath(os.path.join(
-    TESTS_DIR, '..', '..', 'examples', MODEL
+    TESTS_DIR, '..', '..', 'examples', PROJECT
 ))
+sys.path.append(PROJ_PATH)
 
-fid, fn, info = imp.find_module('models',
-                                [os.path.join(PROJ_PATH, MODEL.lower())])
-pvpower_models = imp.load_module('models', fid, fn, info)
+fid, fn, info = imp.find_module(MODEL,
+                                [os.path.join(PROJ_PATH, PROJECT.lower())])
+pvpower_models = imp.load_module(MODEL, fid, fn, info)
 fid.close()
