--- examples/PVPower/pvpower/simulations.py
@@ -1,9 +1,0 @@
-"""
-PV Power Demo Simulations
-"""
-
-from circus.core.simulations import Simulation
-
-
-class SAPM(Simulation):
-    pass
